修复迷你运盘和WEB端不能同时登陆的BUG

+++ protected/extensions/oauth2/server/pdo/lib/PDOOAuth2.php
@@ -194,8 +194,12 @@ class PDOOAuth2 extends OAuth2 {
      */
     protected function getToken($clientId, $scope, $deviceId){
         //如是网页版，则把前一个Token删除，确保系统只有一个用户登录网页版
+        //如果是在pc客户端解锁页面的时候，不用删除设备
         if(MiniHttp::clientIsBrowser()){
-            MiniToken::getInstance()->deleteByDeviceId($deviceId);
+            $source  = $_REQUEST['source'];
+            if(empty($source) || $source!="unlock"){
+                MiniToken::getInstance()->deleteByDeviceId($deviceId);
+            }
         }
         $tokenOauth = MiniToken::getInstance()->getToken4Login($clientId,$deviceId);
         if (!isset($tokenOauth)) {
